api/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
Babel==2.9.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
docutils==0.17.1
esbonio==0.9.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
mysqlclient==2.1.0
packaging==21.3
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pygls==0.11.3
Pygments==2.11.2
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
pyspellchecker==0.6.3
pytz==2021.3
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
toml==0.10.2
typeguard==2.13.3
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0
